Move SnapshotKind to agave-snapshots
