+++ runtime/src/snapshot_package.rs
@@ -2,7 +2,7 @@
 use solana_hash::Hash;
 use {
     crate::bank::{Bank, BankFieldsToSerialize, BankHashStats, BankSlotDelta},
-    agave_snapshots::snapshot_hash::SnapshotHash,
+    agave_snapshots::{snapshot_hash::SnapshotHash, SnapshotKind},
     solana_accounts_db::accounts_db::AccountStorageEntry,
     solana_clock::Slot,
     std::{
@@ -94,22 +94,5 @@ impl std::fmt::Debug for SnapshotPackage {
             .field("slot", &self.slot)
             .field("block_height", &self.block_height)
             .finish_non_exhaustive()
-    }
-}
-
-/// Snapshots come in two kinds, Full and Incremental.  The IncrementalSnapshot has a Slot field,
-/// which is the incremental snapshot base slot.
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum SnapshotKind {
-    FullSnapshot,
-    IncrementalSnapshot(Slot),
-}
-
-impl SnapshotKind {
-    pub fn is_full_snapshot(&self) -> bool {
-        matches!(self, SnapshotKind::FullSnapshot)
-    }
-    pub fn is_incremental_snapshot(&self) -> bool {
-        matches!(self, SnapshotKind::IncrementalSnapshot(_))
     }
 }
